+++ apps/frontend/src/components/DevicesTable.tsx
@@ -266,7 +266,8 @@ function DevicesTable({ filterPlugins, filterDevices }: DevicesTableProps): Reac
       setDeviceName(null);
       return;
     }
-    if (debug || localDebug) console.log(`DevicesTable handleDeviceClick: selected device "${row.name}" with uniqueId "${row.uniqueId}", plugin "${row.pluginName}", endpoint "${row.endpoint}"`);
+    if (debug || localDebug)
+      console.log(`DevicesTable handleDeviceClick: selected device "${row.name}" with uniqueId "${row.uniqueId}", plugin "${row.pluginName}", endpoint "${row.endpoint}"`);
     setSelectedDeviceUniqueId(row.uniqueId);
     setPluginName(row.pluginName);
     setEndpoint(row.endpoint ? row.endpoint.toString() : null);